フォームのバリデーション追加、パスワードの表示/非表示切り替え機能実装
