+++ app/ui/login.tsx
@@ -1,7 +1,10 @@
 import { useFormState } from "react-dom";
+import { useState } from "react";
 import SubmitButton from "./submitButton";
-import { login } from "@/lib/firebase/client";
+import { login } from "@/lib/authentication";
 import { useRouter } from "next/navigation";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faEye, faEyeSlash } from "@fortawesome/free-solid-svg-icons";
 
 const initialState = {
   message: "",
@@ -9,7 +12,12 @@ const initialState = {
 
 export default function LoginComponent() {
   const [error, action] = useFormState(login, initialState);
+  const [isPasswordView, setIsPasswordView] = useState(false);
   const router = useRouter();
+
+  const togglePassword = () => {
+    setIsPasswordView((prevState) => !prevState);
+  };
 
   return (
     <main className="flex flex-col items-center justify-center min-h-screen py-2">
@@ -17,30 +25,42 @@ export default function LoginComponent() {
       <form action={action} className="w-full max-w-xs space-y-4">
         <div className="flex flex-col">
           <label htmlFor="email" className="mb-2">
-            Email:
+            メールアドレス:
           </label>
           <input
             id="email"
             type="email"
             name="email"
-            placeholder="Enter your email address"
+            placeholder="メールアドレスを入力"
             required
             className="appearance-none rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
           />
         </div>
         <div className="flex flex-col">
           <label htmlFor="password" className="mb-2">
-            Password:
+            パスワード:
           </label>
-          <input
-            id="password"
-            type="password"
-            name="password"
-            placeholder="Enter your password"
-            required
-            minLength={6}
-            className="appearance-none rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-          />
+          <div className="relative">
+            <input
+              id="password"
+              type={isPasswordView ? "text" : "password"}
+              name="password"
+              placeholder="パスワードを入力"
+              required
+              minLength={6}
+              className="appearance-none rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+            />
+            <span
+              onClick={togglePassword}
+              className="absolute right-0 top-0 mt-2 mr-2 cursor-pointer"
+            >
+              {isPasswordView ? (
+                <FontAwesomeIcon icon={faEye} />
+              ) : (
+                <FontAwesomeIcon icon={faEyeSlash} />
+              )}
+            </span>
+          </div>
         </div>
         <div className="flex flex-col items-center justify-center">
           <SubmitButton title="ログイン" />
