+++ app/ui/signup.tsx
@@ -1,7 +1,10 @@
 import { useFormState } from "react-dom";
+import { useState } from "react";
 import SubmitButton from "./submitButton";
-import { createUser } from "@/lib/firebase/client";
+import { createUser } from "@/lib/authentication";
 import { useRouter } from "next/navigation";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faEye, faEyeSlash } from "@fortawesome/free-solid-svg-icons";
 
 const initialState = {
   message: "",
@@ -9,6 +12,8 @@ const initialState = {
 
 export default function SignUpComponent() {
   const router = useRouter();
+  const [isPasswordView, setIsPasswordView] = useState(false);
+  const [isPasswordConfirmView, setIsPasswordConfirmView] = useState(false);
   const [error, action] = useFormState(createUser, initialState);
 
   return (
@@ -17,30 +22,70 @@ export default function SignUpComponent() {
       <form action={action} className="w-full max-w-xs space-y-4">
         <div className="flex flex-col">
           <label htmlFor="email" className="mb-2">
-            Email:
+            メールアドレス:
           </label>
           <input
             id="email"
             type="email"
             name="email"
-            placeholder="Enter your email address"
+            placeholder="メールアドレスを入力"
             required
             className="appearance-none rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
           />
         </div>
         <div className="flex flex-col">
           <label htmlFor="password" className="mb-2">
-            Password:
+            パスワード:
           </label>
-          <input
-            id="password"
-            type="password"
-            name="password"
-            placeholder="Enter your password"
-            required
-            minLength={6}
-            className="appearance-none rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-          />
+          <div className="relative">
+            <input
+              id="password"
+              type={isPasswordView ? "text" : "password"}
+              name="password"
+              placeholder="パスワードを入力"
+              required
+              minLength={6}
+              className="appearance-none rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+            />
+            <span
+              onClick={() => setIsPasswordView((prevState) => !prevState)}
+              className="absolute right-0 top-0 mt-2 mr-2 cursor-pointer"
+            >
+              {isPasswordView ? (
+                <FontAwesomeIcon icon={faEye} />
+              ) : (
+                <FontAwesomeIcon icon={faEyeSlash} />
+              )}
+            </span>
+          </div>
+        </div>
+        <div className="flex flex-col">
+          <label htmlFor="password" className="mb-2">
+            パスワード(確認用):
+          </label>
+          <div className="relative">
+            <input
+              id="passwordConfirm"
+              type={isPasswordConfirmView ? "text" : "password"}
+              name="passwordConfirm"
+              placeholder="パスワードを入力(確認用)"
+              required
+              minLength={6}
+              className="appearance-none rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+            />
+            <span
+              onClick={() =>
+                setIsPasswordConfirmView((prevState) => !prevState)
+              }
+              className="absolute right-0 top-0 mt-2 mr-2 cursor-pointer"
+            >
+              {isPasswordConfirmView ? (
+                <FontAwesomeIcon icon={faEye} />
+              ) : (
+                <FontAwesomeIcon icon={faEyeSlash} />
+              )}
+            </span>
+          </div>
         </div>
         <div className="flex flex-col items-center justify-center">
           <SubmitButton title="新規登録" />
